% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ellipsoid_methods.R
\name{add_ellipsoid2persp}
\alias{add_ellipsoid2persp}
\title{add ellipse to persp plot}
\usage{
add_ellipsoid2persp(x, N = 2, colmap = values2colors, pmat, ...)
}
\description{
add ellipse to persp plot
}
