% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnangle.R
\name{angles.cdf}
\alias{angles.cdf}
\title{Cumulative distribution function and pointwise CI}
\usage{
angles.cdf(angles, n = 50, ...)
}
\arguments{
\item{angles}{angles vector, from 2D pattern}

\item{n}{grid resolution}

\item{...}{ignored}
}
\description{
Cumulative distribution function and pointwise CI
}

