% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle-vector-methods.R
\name{angles.flower}
\alias{angles.flower}
\title{Plot a rose of directions for a set of angles}
\usage{
angles.flower(angles, binned, span = 0, k = 25, nums = FALSE,
  ci = FALSE, ...)
}
\arguments{
\item{angles}{observed angles , -pi to pi}

\item{binned}{binning of angles. this or angles needed.}

\item{span}{if >0, span is passed to loess for smoothing and the smoothed values are added to the plot}

\item{k}{number of bins if binning is done for angles}

\item{nums}{plot the values}

\item{ci}{Plot approximate pointwise 95\% confidence interval}

\item{...}{passed to plot (e.g. main)}
}
\description{
Useful for checking e.g. nnangles- or fry_points- outputs.
}

