% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nnangle.R
\name{angles.smooth}
\alias{angles.smooth}
\title{smoothen an histogram of angles}
\usage{
angles.smooth(binned, ...)
}
\arguments{
\item{binned}{binning of angles}

\item{...}{passed to loess}
}
\description{
smoothen an histogram of angles
}

