% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-fast.R
\name{anisotropy_abs_fast}
\alias{anisotropy_abs_fast}
\title{Fast absolute anisotropy  statistic}
\usage{
anisotropy_abs_fast(f, r, eps, antipodal, power)
}
\arguments{
\item{f}{Fry points, border corrected}

\item{r}{range of integration}

\item{eps}{width of cone angle}

\item{antipodal}{If TRUE, antipodes are equated}
}
\description{
Fast absolute anisotropy  statistic
}

