% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{anisotropy_profile}
\alias{anisotropy_profile}
\title{Anisotropy profile assuming compression in one dimension}
\usage{
anisotropy_profile(x, i = 1, cvec, verb = TRUE, ...)
}
\arguments{
\item{x}{point pattern, list(x=coords, bbox=bounding-box)}

\item{i}{direction of compression.}

\item{cvec}{vector of compressions to go through}

\item{...}{passed on to anistropy.abs}
}
\description{
Given a direction i, compute the anisotropy profile for set a compressions.
}
