% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-fast.R
\name{anisotropy_profile_fast}
\alias{anisotropy_profile_fast}
\title{Anisotropy profile for given diagonal transformations}
\usage{
anisotropy_profile_fast(x, grid, r, verb = FALSE, eps = pi/4,
  border = TRUE, antipodal = TRUE, power = 2, ...)
}
\arguments{
\item{x}{point pattern}

\item{grid}{p x dim matrix giving the diagonals of the candidate transformations}

\item{r}{vector of range values to sum over ('integration nodes')}

\item{verb}{verbose?}

\item{eps}{Sector angle}

\item{border}{Border correction? Should use this unless the data is large.}

\item{antipodal}{Use antipodally symmetric estimation of sector-K?}

\item{power}{Power of the differences, absolute value for power=1.}
}
\description{
Compares the directed sector-K functions along main axes.
}
\details{
For each v in grid, transform x -> T(v)x and summarize anisotropy by
absolute integral difference in x-y-z-directed K-functions.

The transformations T(v) are given by diagonal vector v, rows of parameter \code{grid}.

We assume that no rotation is present, or that rotation is already corrected for, so that
squared difference sums of the directed sector-K functions along main axes should indicate
the optimal v.
}
