% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ellipsoid_projections.R
\name{area_of_2d_polygon}
\alias{area_of_2d_polygon}
\title{Compute the area of 2D polygon given by ordered border points}
\usage{
area_of_2d_polygon(px)
}
\arguments{
\item{px}{polygon coordinates, open ended counter clockwise}
}
\description{
So called shoelace method.
}
