% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbpoly.R
\name{as_bbpoly}
\alias{as_bbpoly}
\title{Turn something into a bbpoloy}
\usage{
as_bbpoly(x)
}
\arguments{
\item{x}{input}
}
\description{
Turn something into a bbpoloy
}
\details{
understands: matrix (bbox); bbquad; list with vertices, egdes and faces in 3d
}
