% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ellipsoid_methods.R
\name{as_ellipsoid}
\alias{as_ellipsoid}
\title{Create an ellipsoid object}
\usage{
as_ellipsoid(semi_axes = c(1, 1, 1), R = diag(0, 3), center = c(0, 0, 0))
}
\arguments{
\item{semi_axes}{semi_axes lengths}

\item{R}{rotation matrix}

\item{center}{center coordinates}
}
\description{
Create an ellipsoid object
}
