% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_sphere.R
\name{aziinc2rotationMatrix}
\alias{aziinc2rotationMatrix}
\title{Rotation using azimuth-inclination.}
\usage{
aziinc2rotationMatrix(rot_ai = c(0, 0))
}
\arguments{
\item{rot_ai}{angles}
}
\description{
Rotation using azimuth-inclination.
}
