% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartletts_spectral_density.R
\name{bartletts_spectral_density}
\alias{bartletts_spectral_density}
\title{Bartlett's spectral density}
\usage{
bartletts_spectral_density(x, k = -16:16, std = FALSE, ...)
}
\arguments{
\item{x}{point pattern, list of $x and $bbox}

\item{k}{The frequencies in integers from -n to n. See details.}

\item{std}{scale to unit window?}

\item{...}{ignored}
}
\description{
Periodogram estimator
}
\details{
We estimate the spectral density using a periodogram as suggested by Bartlett 1964.
The frequencies that the spectrum is estimated are omega = 2*pi*k/n, 
where k is a d-dimensional signed integer, (=integer frequency).
The k as input therefore should be either 1) a vector such as -16:16 (default) 
which will be expanded to integer frequencies by taking all pairs using expand.grid or 
2) column matrix of dimension m x d, each row giving one integer frequency. 

No edge correction is applied. 

If std = TRUE, the pattern is first scaled to unit cube.
}
