% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_methods.R
\name{bbox_make}
\alias{bbox_make}
\title{compute the bounding box for coordinates}
\usage{
bbox_make(x, expand = TRUE)
}
\arguments{
\item{x}{coordinate matrix}

\item{expand}{expand a bit? (ripras)}
}
\description{
compute the bounding box for coordinates
}

