% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbpoly.R
\name{bbpoly_affine}
\alias{bbpoly_affine}
\title{Affine transformation of a polytope}
\usage{
bbpoly_affine(bbpoly, A, s = c(0, 0, 0), center_before_A = FALSE)
}
\arguments{
\item{bbpoly}{bbpoly object}

\item{A}{linear transformation}

\item{center_before_A}{Center the polygon to mass centrum before applying A? Reversed after A and before shift.}
}
\description{
Affine transformation of a polytope
}

