% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbpoly-computations.R
\name{bbpoly_correct_faces}
\alias{bbpoly_correct_faces}
\title{Check and correct for orientation of faces}
\usage{
bbpoly_correct_faces(x)
}
\arguments{
\item{x}{bbpoly object in 3D}
}
\description{
Assumes faces with vertices on the same plane
}
