% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbpoly.R
\name{bbpoly_planes}
\alias{bbpoly_planes}
\title{Surface planes of bbpoly}
\usage{
bbpoly_planes(x)
}
\arguments{
\item{x}{bbpoly object}
}
\description{
Surface planes of bbpoly
}
\details{
Returns the normals of the edges (2d) or faces (3d). Assumes  that all vertices of each face are on the same plane.
}
