% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbpoly-computations.R
\name{bbpoly_simplices}
\alias{bbpoly_simplices}
\title{Split a polytope to simplices.}
\usage{
bbpoly_simplices(x, as.bbpoly = FALSE)
}
\arguments{
\item{x}{bbpoly object}

\item{as.bbpoly}{Return the simplices as bbpoly-objects?}
}
\description{
Split a polytope to simplices.
}
\details{
Assumes convexity, will not check.
}

