% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbpoly-computations.R
\name{bbpoly_volume}
\alias{bbpoly_volume}
\title{Volume of a bbpoly}
\usage{
bbpoly_volume(x)
}
\arguments{
\item{x}{bbpoly object}
}
\description{
Volume of a bbpoly
}
\details{
splits the bbpoly into simplices and sums  their volume.
}

