% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbquad.R
\name{bbquad_affine}
\alias{bbquad_affine}
\title{Affine transformation of a bbquad}
\usage{
bbquad_affine(x, A, s = c(0, 0, 0))
}
\arguments{
\item{x}{bbquad object}

\item{A}{transformation matrix}

\item{s}{shift vector}
}
\description{
Affine transformation of a bbquad
}
