% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbquad.R
\name{bbquad_default}
\alias{bbquad_default}
\title{Regular bounding box as bounding box of quadrilaterals}
\usage{
bbquad_default(xl = c(-0.5, 0.5), yl = xl, zl = xl)
}
\description{
Regular bounding box as bounding box of quadrilaterals
}

