% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbquad.R
\name{bbquad_distance}
\alias{bbquad_distance}
\title{Distance from points to bbquad walls}
\usage{
bbquad_distance(x, b, warn = TRUE)
}
\arguments{
\item{x}{n x 3 matrix of locations}

\item{b}{quadrilateral bounding box}

\item{warn}{Emit a warning for points outside box? (default: TRUE)

Note: Returned distance will be negative if point outside the box.}
}
\description{
Distance from points to bbquad walls
}
