% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbquad.R
\name{bbquad_simplices}
\alias{bbquad_simplices}
\title{Triangulate a bbquad}
\usage{
bbquad_simplices(x)
}
\arguments{
\item{x}{bbquad object}
}
\description{
Due to assuming quadrilateral polytope, we get 2 simplices (triangles/tetrahedrons)
}
