% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbquad.R
\name{bbquad_volume}
\alias{bbquad_volume}
\title{Volume of a bbquad}
\usage{
bbquad_volume(x)
}
\arguments{
\item{x}{bbquad-object}
}
\description{
Volume of a bbquad
}
\details{
Will split the bbquad into simplices and return the sum of simplex volumes.
}
