% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_directions.R
\name{check_directions}
\alias{check_directions}
\title{Sector computation directions}
\usage{
check_directions(theta, epsilon, unit, dim = 2, antipode = FALSE, n_dir = 12)
}
\arguments{
\item{theta}{angles}

\item{epsilon}{opening}

\item{unit}{unit vecs}

\item{dim}{dimension}

\item{antipode}{include antipodal directions}

\item{n_dir}{resolution of generated directions}
}
\description{
Internal use only. Check input or generate directions.
}
