% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ellipsoid_rellipsoid.R
\name{check_vector_passes_triangle}
\alias{check_vector_passes_triangle}
\title{Check if a vector (or vectors), multiplied by inf, go(es) through a triangle in 3D}
\usage{
check_vector_passes_triangle(u, triangle)
}
\arguments{
\item{u}{unit vector}

\item{triangle}{triangle coordinates}
}
\description{
Check if a vector (or vectors), multiplied by inf, go(es) through a triangle in 3D
}
