% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ellipsoid_confint.R
\name{confint.ellipsoid_old}
\alias{confint.ellipsoid_old}
\title{Confidence interval for ellipsoid parameters using simulation}
\usage{
\method{confint}{ellipsoid_old}(x, fun, nsim = 1000, probs = c(0.025, 0.975), ...)
}
\arguments{
\item{x}{Fitted ellipsoid}

\item{fun}{Optional contrast function. See ellipse_contrast_2d for an example.}

\item{probs}{The quantiles for the CI. Default is c(0.025, 0.975)}

\item{tol}{tolerance for absolute  deviation in the ||beta||=1 constraint.}
}
\description{
Assuming normality of errors.
}
