% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint-fry-ellipsoids.R
\name{confint.fryellipsoids}
\alias{confint.fryellipsoids}
\title{Confidence intervals for Fry ellipsoids}
\usage{
\method{confint}{fryellipsoids}(object, parm, level = 0.95, fun, ...)
}
\arguments{
\item{parm}{not used}

\item{level}{not used}

\item{fun}{contrast function, passed on to confint.ellipsoid}

\item{...}{Passed to confint.ellipsoid}

\item{x}{Fitted Fry-ellipsoids object}
}
\description{
Confidence intervals for Fry ellipsoids
}
\details{
See the 'confint.ellipsoid' in 'ellipsoid' for details of ...
}
