% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_methods.R
\name{coord_affine}
\alias{coord_affine}
\title{Affine transfomration of coordinates}
\usage{
coord_affine(x, A, s = c(0, 0, 0))
}
\arguments{
\item{x}{coordinates, one row per coordinate}

\item{A}{matrix}

\item{s}{shift}
}
\description{
Affine transfomration of coordinates
}

