% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalDist_Konig_Schmidt.R
\name{dDirectional}
\alias{dDirectional}
\title{Nearest neighbour directional distribution}
\usage{
dDirectional(x, r, theta, epsilon, n_dir = 10)
}
\arguments{
\item{x}{point pattern, $x coords $bbox bounding box}

\item{r}{Ranges}

\item{theta}{Direction sector central angles, list of ncol(x$x)-components.}

\item{epsilon}{Direction sector width angles}

\item{n_dir}{If theta not given, greate a grid with this resolution.}
}
\description{
Estimate the directional distribution of Konig&Schmidt 1992
with (their notation) 0-nearest neighbour i.e. nearest neighbour.
}
\details{
Directions are angle to positive x-axis (2D) or azimuth, inclination (3D). 
Theta should be a list of length 1 (2D) and 2 (3D) defining a grid of these angles.
}

