% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodogram.R
\name{DFT}
\alias{DFT}
\title{Compute the Discrete Fourier Transform of a point pattern}
\usage{
DFT(x, m)
}
\arguments{
\item{x}{point pattern}

\item{m}{maximum steps}
}
\description{
2D only
}
\details{
The pattern will be rescaled to unit square.
}
