% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/directed_geom.R
\name{directed_geom}
\alias{directed_geom}
\title{Directed geometric graph}
\usage{
directed_geom(x, u, theta, r, from, to, double = TRUE, pregraph)
}
\arguments{
\item{x}{matrix of location coordinates}

\item{u}{unit vector of direction}

\item{theta}{angle for the directed cone}

\item{r}{radius for neighbourhood}

\item{from}{default 1:nrow(x), indices of points from which to compute edges}

\item{to}{default 1:nrow(x), indices of points which account as potential neighbours}

\item{double}{use the double cone, i.e. look also in direction -u?}

\item{pregraph}{A list-of-neighbour indices where to look for potential neighbours}
}
\value{
Neighbour index vectors as a list.
}
\description{
Compute the geometric graph \deqn{x_i ~ x_j \Leftrightarrow ||x_i-x_j|| < r \& -\theta < angle(x_j-x_i) < \theta}{x1 ~ x2 <=> ||x1-x2||< r & -theta < angle(x2-x1, u) < theta}
}
\details{
If pregraph is given, "to" is ignored.
}

