% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbquad.R
\name{dist_point_to_plane}
\alias{dist_point_to_plane}
\title{Point to plane distance}
\usage{
dist_point_to_plane(x, pl, absit = TRUE)
}
\arguments{
\item{x}{vector of 3 coordinates}

\item{pl}{plane in norm-point format}

\item{absit}{take abs? Otherwise sign gives side}
}
\description{
Point to plane distance
}
