% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.arc.R
\name{draw.arc}
\alias{draw.arc}
\title{Draw arc}
\usage{
draw.arc(x = 1, y = NULL, radius = 1, angle1 = deg1 * pi/180,
  angle2 = deg2 * pi/180, deg1 = 0, deg2 = 45, n = 0.05,
  col = NA, lwd = NA, ...)
}
\arguments{
\item{x}{x coordinate of center. Scalar or vector.
...}
}
\description{
(From plotrix) Draw one or more arcs using classic graphics.
}
