% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ellipsoid_contrasts.R
\name{ellipse_contrast_2d}
\alias{ellipse_contrast_2d}
\title{Testing contrast for ellipses}
\usage{
ellipse_contrast_2d(els, out = TRUE, range = 1.85)
}
\arguments{
\item{out}{do outlier removal, the quadratic simulations is unstable.}

\item{range}{range of outlier removal. See boxplot for details of the formula.}
}
\description{
Give this to confint.ellipsoid to compute the contrast for
 testing semi_a - semi_b = 0
}
