% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ellipsoid_methods.R
\name{ellipse_form}
\alias{ellipse_form}
\title{###################################################################
Ellipse center and matrix from general parameter form}
\usage{
ellipse_form(beta, d, check = FALSE)
}
\arguments{
\item{beta}{OLS estimates}

\item{d}{dimension}

\item{check}{Check for definiteness?}
}
\description{
###################################################################
Ellipse center and matrix from general parameter form
}
