% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ellipsoid_methods.R
\name{ellipse_solve_rota}
\alias{ellipse_solve_rota}
\title{########################################################################
Solve rotation and semi-axes from general transform}
\usage{
ellipse_solve_rota(A, eps = 0)
}
\arguments{
\item{A}{the trasformation matrix in an ellipsoid equation}

\item{eps}{Inflate diagonal by this factor, to avoid numerical problems.}
}
\description{
########################################################################
Solve rotation and semi-axes from general transform
}
\details{
does not work with improper A, will return inf long axes.
}
