% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ellipsoid_ALS.R
\name{ellipsoid_ALS}
\alias{ellipsoid_ALS}
\title{ALS fitting of ellipsoid}
\usage{
ellipsoid_ALS(x, s2, ..., just_U = FALSE)
}
\arguments{
\item{x}{Point locations, n x d matrix}

\item{s2}{Error variance, or a vector over which to optimize.}

\item{...}{ignored}

\item{just_U}{For internal use: Return just smallest eigenvector and bhat.}
}
\description{
Adjusted LS fitting of ellipsoids, by Kukush et al. 2003.
}
\details{
Not really optimised yet. Good grid for s2 optimisation is something like
seq(s2 * 0.1, s2, length=20)
where s2 is the OLS estimate (access via $ols_fit$s2).
}
\references{
Kukush et al (2003): Consistent estimation in an implicit quadratic measurement error model,
Computational Stat & Data Analysis
}
