% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ellipsoid_methods.R
\name{ellipsoid_from_beta}
\alias{ellipsoid_from_beta}
\title{convert beta vector to ellipsoid object}
\usage{
ellipsoid_from_beta(beta, d, ...)
}
\arguments{
\item{beta}{beta vector, the coefficients in quadratic form}

\item{d}{dimension}

\item{...}{passed on to 'ellipse_solve_rota'}
}
\description{
convert beta vector to ellipsoid object
}
