% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ellipsoid_OLS.R
\name{ellipsoid_OLS}
\alias{ellipsoid_OLS}
\title{Ordinary Least Squares Ellipsoid}
\usage{
ellipsoid_OLS(x, origin = FALSE)
}
\arguments{
\item{x}{point coordinate matrix}

\item{origin}{Fix center of the ellipsoid to be the origin?}
}
\description{
OLS fitting of an ellipsoid to a d-dimensional point cloud
}
\details{
Assumes the normal error model.
}
