% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ellipsoid_methods.R
\name{ellipsoid_shape}
\alias{ellipsoid_shape}
\title{Ellipsoid shape for 3d plotting}
\usage{
ellipsoid_shape(N = 2, axes = c(1, 1, 1), R = NULL, center = c(0, 0, 0))
}
\arguments{
\item{R}{Rotation matrix.}
}
\description{
Refine an icosahedron, then transform
}
