% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_sphere.R
\name{Euler2rotationMatrix}
\alias{Euler2rotationMatrix}
\title{Euler axis-angle to Rotation matrix}
\usage{
Euler2rotationMatrix(v, proper = TRUE)
}
\arguments{
\item{v}{euler vector}

\item{proper}{proper}
}
\description{
Euler axis-angle to Rotation matrix
}
