% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_sphere.R
\name{EulerAngles2quaternion}
\alias{EulerAngles2quaternion}
\title{Euler angles to quaternions}
\usage{
EulerAngles2quaternion(angle)
}
\arguments{
\item{angle}{Euler angle vector}
}
\description{
Euler angles to quaternions
}
