% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fest_directional.R
\name{Fest_along_axis}
\alias{Fest_along_axis}
\title{Compute all directed F functions along axis}
\usage{
Fest_along_axis(x, ...)
}
\arguments{
\item{x}{pp, list with $x~coordinates $bbox~bounding box}

\item{...}{parameters except u for Kest_directional}
}
\description{
Compute all directed F functions along axis
}

