% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fest_directional.R
\name{Fest_directional}
\alias{Fest_directional}
\title{Directed F, empty space function}
\usage{
Fest_directional(x, u, theta, r, n)
}
\arguments{
\item{x}{pp, list with $x~coordinates $bbox~bounding box}

\item{u}{unit vector of direction}

\item{theta}{angle for the directed cone}

\item{r}{radius vector at which to estimate}

\item{n}{number of uniform dummies}
}
\description{
We use toroidal
}
