% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods_pcf_directions.R
\name{flower.pcf_directions}
\alias{flower.pcf_directions}
\title{Plot a angluar plot of the pcf_directions, 2D}
\usage{
flower.pcf_directions(x, scale_points = TRUE, col = rainbow, zlim, ...)
}
\description{
Plot a angluar plot of the pcf_directions, 2D
}

