% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fry-ellipsoids-2.R
\name{fry_ellipsoids2}
\alias{fry_ellipsoids2}
\title{Fit second order intensity ellipses or ellipsoids}
\usage{
fry_ellipsoids2(x, nvec = 1:5, r_adjust = 1, nangles, eps = 0,
  cylindric = FALSE, double = FALSE, border = TRUE, origin = TRUE,
  verbose = FALSE, keep_data = FALSE, data, ...)
}
\arguments{
\item{x}{point pattern}

\item{nvec}{which jumps to consider}

\item{r_adjust}{consider only fry points with length r_adjust*max(window side)}

\item{nangles}{If 3d, number of refinements (def=2), in 2d def=20 sectors.}

\item{eps}{add to the sector angle. Results in smoothing if >0 (default=0).}

\item{cylindric}{Use a cylinder instead of a sector}

\item{double}{Should we use both directions of a pair}

\item{border}{Should border effects be cancelled by minus sampling, if TRUE, "double" ignored.}

\item{origin}{Constrain the ellipses' centers to origo.}

\item{keep_data}{Keep the data for each ellipsoid, inside each ellipsoids? Helps with averaging.}

\item{data}{The vectors to which fit ellipses. Default is missing, and we compute Fry-points.}
}
\description{
Difference to v1: Don't use the pseudo-Fry but true Fry.
}

