% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second_order_orientation_density.R
\name{fry_orientation_density}
\alias{fry_orientation_density}
\title{Second order orientation density}
\usage{
fry_orientation_density(x, r1, r2, alpha, bw = 0.35)
}
\arguments{
\item{x}{point pattern}

\item{r1}{lower limit of the ranges}

\item{r2}{upper limit of the ranges}

\item{alpha}{angles to estimate the density at, on range [0, pi]}

\item{bw}{Bandwidth (half-width) for the Epanechnicov kernel.}
}
\description{
As in Illian et al 2008, Ch. 4.5.3.
}
