% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fry_points.R
\name{fry_points}
\alias{fry_points}
\title{Fry points}
\usage{
fry_points(x, double = FALSE, border = TRUE)
}
\arguments{
\item{double}{Should we return both directions of a pair}

\item{border}{Should border effects be cancelled by minus sampling, if TRUE, "double" ignored.}
}
\value{
Returns the lengths and direction unit vectors.
}
\description{
Pairwise difference vectors of x
}
