% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gsector.R
\name{Gsector}
\alias{Gsector}
\title{Directional Nearest neighbour distribution(old)}
\usage{
Gsector(x, r, theta, unit, directions, epsilon, n_dir = 10)
}
\arguments{
\item{x}{point pattern, $x coords $bbox bounding box}

\item{r}{Ranges}

\item{theta}{Direction sector central angles, list of ncol(x$x)-components.}

\item{unit}{Alternative to theta, a matrix of direction vectors.}

\item{epsilon}{Direction sector width angles}

\item{n_dir}{If theta not given, greate a grid with this resolution.}
}
\description{
Estimate the directional nearest neighbour CDF.
}
\details{
This is a directed sector version of the G-function in spatstat.

Directions are angle to positive x-axis (2D) or azimuth, inclination (3D). 
Theta should be a list of length 1 (2D) and 2 (3D) defining a grid of these angles.
}
