% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{intensity_somewhere}
\alias{intensity_somewhere}
\title{Kernel Estimate Intensity at non-Data locations}
\usage{
intensity_somewhere(x, loc, bw, kernel = "gaussian", border = "local")
}
\arguments{
\item{x}{point pattern}

\item{loc}{coordinate matrix for non-data locations at which to estimate the intensity.}

\item{bw}{bandwidth. Gaussian sd=bw, Epanechnicov domain [-bw, bw].}

\item{kernel}{Either 'gaussian' or 'epanenchnikov'}

\item{border}{Border correction to apply.  One of 'none', 'local', 'global'.}

\item{normalise}{renormalise so that inverse sum = volume}

\item{loo}{leave-one-out -estimate?}
}
\description{
Kernel Estimate Intensity at non-Data locations
}
\details{
For border correction, the bounding box of the coordinates of x will be used, 
so at the moment it works only for axis-aligned cuboids.

The kernel will be a product of one dimensional kernels,
so Epanechnikov in many dimensions will not be truly isotropic but prefers diagonal directions.
}
