% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Kdirectional.R
\docType{package}
\name{Kdirectional}
\alias{Kdirectional}
\alias{Kdirectional-package}
\title{Some tools for analysing the isotropy of spatial point patterns}
\description{
See the vignette for an example.
}

