% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Kest_directional.R
\name{Kest_along_axis}
\alias{Kest_along_axis}
\title{Compute all directed K functions along axis}
\usage{
Kest_along_axis(x, ...)
}
\arguments{
\item{x}{pp, list with $x~coordinates $bbox~bounding box}

\item{...}{parameters except u for Kest_directional}
}
\description{
Compute all directed K functions along axis
}

