% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kest_anin.R
\name{Kest_anin}
\alias{Kest_anin}
\title{Inhomogeneous anisotropic K function}
\usage{
Kest_anin(x, u, epsilon, r, lambda = NULL, lambda_h, renormalise = TRUE,
  border = 1, ...)
}
\arguments{
\item{x}{pp, list with $x~coordinates $bbox~bounding box}

\item{u}{unit vector(s) of direction, as row vectors. Default: x and y axis.}

\item{epsilon}{Central half angle for the directed sector/cone (total angle of the rotation cone is 2*epsilon)}

\item{r}{radius vector at which to evaluate K}

\item{lambda}{optional vector of intensity estimates at points}

\item{lambda_h}{if lambda missing, use this bandwidth in a kernel estimate of lambda(x)}

\item{renormalise}{See details.}

\item{border}{Use translation correction? Default=1, yes. Only for cuboidal windows.}

\item{...}{passed on to e.g. \link{intensity_at_points}}
}
\value{
Returns a dataframe.
}
\description{
Estimate a Sector-K function for second order reweighted ("inhomogeneous") pattern.
}
\details{
Computes a second order reweighted version of the Sector-K.

Lambda(x) at points can be given, 
or else it will be estimated using Epanechnikov kernel smoothing. See 

If 'renormalise=TRUE', we normalise the lambda estimate so that sum(1/lambda(x))=|W|. This corresponds in \code{spatstat}'s \code{Kinhom} to setting 'normpower=2'.
}

