% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kest_directional.R
\name{Kest_directional}
\alias{Kest_directional}
\title{Directed K function}
\usage{
Kest_directional(x, u, epsilon, r, ...)
}
\arguments{
\item{x}{pp, list with $x~coordinates $bbox~bounding box}

\item{u}{unit direction(s), if many then one direction per row.}

\item{epsilon}{Central angle for the directed cone (total angle is 2 epsilon)}

\item{r}{radius vector at which to evaluate K}

\item{...}{Passed on to \link{\code{Kest_anin}}}
}
\value{
Returns a dataframe.
}
\description{
We use translation edge correction if we have an axis-oriented bounding box. Otherwise minus-correction.
}
\details{
Compute the sector K function. This version uses the more general anisotropic-inhomonogeneous \link{\code{Kest_anin}}, by setting the intensity = constant.
}

