% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kest_directional.R
\name{Kest_directional_old}
\alias{Kest_directional_old}
\title{Directed K function, old and slow}
\usage{
Kest_directional_old(x, u, epsilon, r, pregraph)
}
\arguments{
\item{x}{pp, list with $x~coordinates $bbox~bounding box}

\item{u}{unit vector of direction}

\item{epsilon}{Central angle for the directed cone (total angle is 2 epsilon)}

\item{r}{radius vector at which to evaluate K}

\item{pregraph}{A super graph for determining neighbourhoods}
}
\value{
Returns a dataframe.
}
\description{
We use translation edge correction.
}
\details{
Compute the sector K function. 

Try \link{Kest_anin} for a bit better version that does also second order inhomogeneous patterns.
}
