% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kest_gaussian.R
\name{Kest_gaussian}
\alias{Kest_gaussian}
\title{Gaussian Weighted Sum of Pairwise Vectors}
\usage{
Kest_gaussian(
  x,
  u,
  kappa,
  r,
  lambda = NULL,
  lambda_h,
  renormalise = TRUE,
  border = 1,
  ...
)
}
\arguments{
\item{x}{pp, list with $x~coordinates $bbox~bounding box}

\item{u}{unit vector(s) of direction, as row vectors. Default: x and y axes, viz. c(1,0) and c(0,1). This gives the major axis of the ellipsoid (direction of largest variance in the Gaussian density).}

\item{kappa}{The ratio of the secondary axis to the main axis going along 'u'.}

\item{r}{radius vector at which to evaluate. Corresponds to the radii of the 95\% quantile in x-axis, before rotation in directions u.}

\item{lambda}{optional vector of intensity estimates at points}

\item{lambda_h}{if lambda missing, use this bandwidth in a kernel estimate of lambda(x)}

\item{renormalise}{See details.}

\item{border}{Use border correction? Default=1, yes. At the moment no other version available!}

\item{...}{passed on to e.g. \link{intensity_at_points}}
}
\value{
Returns a dataframe.
}
\description{
Estimate K function -type summary for second order reweighted ("inhomogeneous") pattern using a Gaussian kernel sitting in the origin of the Fry-plot.
}
\details{
TODO Gaussian kernel sitting at the fry-space, sum over fry-points.
}
