% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnangle.R
\name{knnangle}
\alias{knnangle}
\title{k-th nearest neighbour angles and distances}
\usage{
knnangle(x, k = 1, from, to)
}
\arguments{
\item{x}{Matrix of coordinates, n rows and d columns}

\item{k}{k-th nearest neighbour}

\item{from}{From these indices (default all 1,...,nrow(x))}

\item{to}{To these indices (default all)}
}
\value{
The values in 2d are (azimuth, distance)
The values in 3d are physical coordinate system (azimuth, inclination, distance)
}
\description{
k-th nearest neighbour angles and distances
}
\details{
If 2d, returns the angle in [-pi, pi]. In 3d also the angle from xy-plane in
[0,pi]
}
