% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.test2.R
\name{ks.test2}
\alias{ks.test2}
\title{Kolmogorov-Smirnov test using ecdf's}
\usage{
ks.test2(x, y, n)
}
\arguments{
\item{x}{empirical cdf function, see details}

\item{y}{theoretical cdf function}

\item{n}{Number of 'datapoints'}
}
\description{
Kolmogorov-Smirnov test using ecdf's
}

