% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani-compression-profile.R
\name{lines.anisotropyprofile}
\alias{lines.anisotropyprofile}
\title{Lines method for "anisotropyprofile"}
\usage{
\method{lines}{anisotropyprofile}(x, ..., scale = TRUE)
}
\description{
Lines method for "anisotropyprofile"
}
