% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markcorr_anin.R
\name{markcorr_anin}
\alias{markcorr_anin}
\title{Inhomogeneous anisotropic mark-correlation function, sector version}
\usage{
markcorr_anin(x, marks, u, epsilon, r, lambda = NULL, lambda_h,
  f = function(a, b) a * b, r_h, stoyan = 0.15, renormalise = TRUE,
  bootsize = 1e+05, normaliser = NULL, border = 1, divisor = "d", ...)
}
\arguments{
\item{x}{pp, list with $x~coordinates $bbox~bounding box}

\item{marks}{if x is not marked (x$marks is empty), use these marks}

\item{u}{unit vector(s) of direction, as row vectors. Default: x and y axis.}

\item{epsilon}{Central half angle for the directed sector/cone (total angle of the rotation cone is 2*epsilon)}

\item{r}{radius vector at which to evaluate}

\item{lambda}{optional vector of intensity estimates at points}

\item{lambda_h}{if lambda missing, use this bandwidth in a kernel estimate of lambda(x)}

\item{f}{test function of the form function(m1, m2) ..., returning a vector of length(m1). default: m1*m2}

\item{r_h}{smoothing for range dimension, epanechnikov kernel}

\item{stoyan}{If r_h not given, use r_h=stoyan/lambda^(1/dim). Same as 'stoyan' in spatstat's pcf.}

\item{renormalise}{See details.}

\item{bootsize}{bootstrap size for estimating the normaliser if not given.}

\item{normaliser}{normalising constant under independent marking. If NULL, estimated with bootstrap.}

\item{border}{Use translation correction? Default=1, yes. Only for cuboidal windows.}

\item{divisor}{either "d" or "r". Divide by dist(i,j) ("d") instead of r ("r")?}

\item{...}{passed on to e.g. \link{intensity_at_points}}
}
\value{
Returns a dataframe.
}
\description{
Estimate a sector-mark correlation function for second order reweighted "inhomogeneous" pattern.
}
\details{
Computes a second order reweighted version of the mark correlation function.

lambda(x) at points can be given, 
or else it will be estimated using Epanechnikov kernel smoothing. See 

If 'renormalise=TRUE', we normalise the lambda estimate so that sum(1/lambda(x))=|W|. This corresponds in \code{spatstat}'s \code{Kinhom} to setting 'normpower=2'.
}

