% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani-markcor.R
\name{markcorr_anisotropic}
\alias{markcorr_anisotropic}
\title{Anisotropic mark correlation function, direction vector formulation}
\usage{
markcorr_anisotropic(x, marks = NULL, r, f = function(a, b) a * b,
  directions, bw, adjust = c(1, 1), correction = "translation",
  bootsize = 1e+05, divisor = "d", normaliser = NULL)
}
\arguments{
\item{x}{pp, list with $x~coordinates $bbox~bounding box}

\item{marks}{if x is not marked (x$marks is empty), use these marks}

\item{r}{Evaluate at these lengths.}

\item{f}{test function of the form function(m1, m2) ..., returning a vector of length(m1)}

\item{directions}{Matrix of directions, in unit vectors, one direction per row. Default: along axes.}

\item{bw}{Bandwidths of epanechnicov kernels. Vector of two values, one for ranges and one for angles.}

\item{adjust}{If bw not given, use bw=adjust[1]*0.15/lambda^(1/dim) for range and bw=adjust[2]*0.15*pi for angle}

\item{correction}{"none" or "translation". Translation only for rectangular box.}

\item{bootsize}{bootstrap size for estimating the normaliser if not given.}

\item{divisor}{either "d" or "r". Divide by dist(i,j) ("d") instead of r ("r")?}

\item{normaliser}{normalising constant under independent marking. If NULL, estimated with bootstrap.}
}
\description{
Estimate the anisotropic mark correlation function (2d and 3d), as defined in Stoyan 1991, f. 5.4. (without the additional 'a'-marks)
}
\details{
Default bandwidth: bw=adjust*0.15/lambda^(1/dim) for range and bw=adjust*0.15*pi for angle.
}
