% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ellipsoid_mean.R
\name{mean_ellipse}
\alias{mean_ellipse}
\title{Mean of ellipses}
\usage{
mean_ellipse(x, weight, ...)
}
\arguments{
\item{x}{a list of ellipses (2d)}

\item{weight}{weighted average weights}
}
\description{
2D ellipsoid averaging.
}
