% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ellipsoid_mean.R
\name{mean_rotation_quat}
\alias{mean_rotation_quat}
\title{Mean ellipsoid rotation using quarternion average}
\usage{
mean_rotation_quat(x, ...)
}
\arguments{
\item{x}{list of ellipsoids}

\item{...}{ignored}
}
\description{
Mean ellipsoid rotation using quarternion average
}
\details{
Average in the quarternion space
}
