% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalDist_Konig_Schmidt.R
\name{nn_directional_angle_distribution}
\alias{nn_directional_angle_distribution}
\title{Nearest neighbour directional distribution (new version)}
\usage{
nn_directional_angle_distribution(x, direction, epsilon, r,
  antipodal = FALSE)
}
\arguments{
\item{x}{point pattern, $x coords $bbox bounding box}

\item{direction}{Unit vector direction}

\item{epsilon}{Direction sector's opening half-angle vector.}

\item{r}{Maximum range for of nn-distances to consider. If missing, use max(nn-distances)}

\item{antipodal}{Make nn-vectors antipodally symmetric?}
}
\description{
Estimate the directional distribution of Konig&Schmidt 1992
with (their notation) 0-nearest neighbour i.e. nearest neighbour, the direction
set being a conical wedge of the unit-sphere.
}
\details{
Compute the probability that a nearest neighbour of a typical point is in direction of the
cone given the distance is less than r. 

Input should be one direction and many epsilons OR equal amount of directions and epsilons.

Not using a double cone.
}
