% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gsector.R
\name{nn_directional_distance_distribution}
\alias{nn_directional_distance_distribution}
\title{Directional nearest neighbour distribution}
\usage{
nn_directional_distance_distribution(x, r, direction, epsilon,
  type = "local", antipodal = TRUE)
}
\arguments{
\item{x}{point pattern}

\item{r}{ranges to consider, increasing vector}

\item{direction}{unit vector(s) of directions of cones}

\item{epsilon}{Cone central half-angle}

\item{type}{"global" or "local" (default).}

\item{antipodal}{Use antipodal symmetry (=double cone)? Default: FALSE}
}
\description{
Global/local directed nearest neighbour distance distribution, conical wedge defining direction set.
}
