% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gsector.R
\name{nn_directional_distance_distribution_global}
\alias{nn_directional_distance_distribution_global}
\title{Directional nearest neighbour distribution, global version}
\usage{
nn_directional_distance_distribution_global(x, u, epsilon, r, antipodal)
}
\description{
Directional nearest neighbour distribution, global version
}
\details{
Called by 'nn_directional_distance_distribution' function.
}
