% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gsector.R
\name{nn_directional_distance_distribution_local}
\alias{nn_directional_distance_distribution_local}
\title{Directional nearest neighbour distribution, local version}
\usage{
nn_directional_distance_distribution_local(x, u, epsilon, r, antipodal)
}
\description{
Directional nearest neighbour distribution, local version
}
\details{
Called by 'nn_directional_distance_distribution' function.
}
