% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnangle-density.R
\name{nn_orientation_density}
\alias{nn_orientation_density}
\title{Nearest neighbour angle density}
\usage{
nn_orientation_density(x, bw = 0.35, antipodal = TRUE, avec, ..., data,
  justData = FALSE)
}
\arguments{
\item{x}{point pattern}

\item{bw}{Epanechnikov kernel bandwidth}

\item{antipodal}{Work with angles only on the 0 - pi range? If false, work on 0 - 2*pi.}

\item{avec}{vector of angles to estimate the density at}

\item{...}{ignored}

\item{data}{alternative to x (expert use only)}

\item{justData}{Just return the components for calculation?}
}
\description{
2D only.
}
