% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnangle.R
\name{nnangle}
\alias{nnangle}
\title{nearest neighbour angles and distances}
\usage{
nnangle(x, from, to)
}
\arguments{
\item{x}{Matrix of coordinates, n rows and d columns}
}
\value{
The values in 2d are (azimuth, distance).
The values in 3d are physical coordinate system (azimuth, inclination, distance)
}
\description{
nearest neighbour angles and distances
}
\details{
If 2d, returns the angle in [-pi,pi]. In 3d also the angle from xy-plane in
[0,pi]
}

