% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohser_stoyan_K.R
\name{osK}
\alias{osK}
\title{Ohser-Stoyan K-function}
\usage{
osK(x, r, theta, ...)
}
\arguments{
\item{x}{point pattern, $x and $bbox}

\item{r}{vector of ranges}

\item{theta}{vector of angles. list with two components if 3D}
}
\description{
Ohser-Stoyan K-function
}

