% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_vectors.R
\name{pairwise_vectors}
\alias{pairwise_vectors}
\title{Compute pairwise distances and angles}
\usage{
pairwise_vectors(x, from, to, asMatrix = FALSE, as.xyz = FALSE)
}
\arguments{
\item{x}{matrix of coordinates.}

\item{from}{Indices from which to compute}

\item{to}{Indices to which to compute}

\item{asMatrix}{return n x n matrices}

\item{as.xyz}{return as xyz-coordinates instead of polar/spherical}
}
\description{
Computes pairwise vectors of a 2d or 3d point pattern. Returns 
circular- or spherical coordinates, depending on the dimension.
}
\details{
Default is to return the upper triangle (without diagonal) of 
pairwise distance and angle matrices.

Result will have 2 or 3 columns, where first is the distance
and 2(:3) are angle(s).
}
