% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_marks.R
\name{parse_marks}
\alias{parse_marks}
\title{Parse input for marks}
\usage{
parse_marks(x, marks)
}
\arguments{
\item{x}{Input object}

\item{marks}{potential marks in case x has none.}
}
\description{
Parse input for marks
}
