% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf_anin_fry.R
\name{pcf_anin_fry}
\alias{pcf_anin_fry}
\title{Inhomogeneous anisotropic pcf function, non-spherical version}
\usage{
pcf_anin_fry(
  x,
  u,
  r,
  lambda = NULL,
  lambda_h,
  bw,
  stoyan = 0.35,
  renormalise = TRUE,
  border = 1,
  ...
)
}
\arguments{
\item{x}{pp, list with $x~coordinates $bbox~bounding box}

\item{u}{unit vector(s) of direction, as row vectors. Default: x and y axis.}

\item{r}{radius vector at which to evaluate the function}

\item{lambda}{optional vector of intensity estimates at points}

\item{lambda_h}{if lambda missing, use this bandwidth in a kernel estimate of lambda(x)}

\item{bw}{smoothing bandwidth for multi-dimensional epanechnikov kernel.}

\item{stoyan}{If r_h not given, use bw=stoyan/lambda^(1/dim). Same as 'stoyan' in spatstat's pcf.}

\item{renormalise}{See details.}

\item{border}{Use translation correction? Default=1, yes. Only for cuboidal windows.}

\item{...}{passed on to e.g. \link{intensity_at_points}}
}
\value{
Returns a dataframe.
}
\description{
Estimate the anisotropic pcf function for second order reweighted ("inhomogeneous") pattern.
}
\details{
Computes a second order reweighted version of the anisotropic pcf. Essentially the estimate of intensity of pairwise vectors at locations
outer(r, u). We will use antipodal estimation, so directions should not include antipodal pairs.

Lambda(x) at points can be given, 
or else it will be estimated using Epanechnikov kernel smoothing. See 

If 'renormalise=TRUE', we normalise the lambda estimate so that sum(1/lambda(x))=|W|. This corresponds in \code{spatstat}'s \code{Kinhom} to setting 'normpower=2'.
}
\seealso{
\code{\link{pcf_anin}}
}
