% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf_anisotropic.R
\name{pcf_anisotropic}
\alias{pcf_anisotropic}
\title{Anisotropic pair correlation function}
\usage{
pcf_anisotropic(x, r, u, h, stoyan = 0.15, lambda, lambda_h,
  renormalise = TRUE, border = 1, divisor = "d", ...)
}
\arguments{
\item{x}{pp, list with $x~coordinates $bbox~bounding box}

\item{r}{radius vector at which to estimate}

\item{u}{direction to be converted to angles. Use \code{\link{angle_2_unit}} to transform from angles.}

\item{h}{half-widths of epanechnicov kernels, vector of two values, one for ranges and and one for angles.}

\item{stoyan}{If h not given, use h=c( stoyan/lambda^(1/dim), stoyan*pi) (cf. \code{pcf.ppp}-function in \code{spatstat}-package).}

\item{lambda}{optional vector of intensity estimates at points}

\item{lambda_h}{if lambda missing and lambda_h is given, use this bandwidth in a kernel estimate of lambda(x). Otherwise lambda is set to constant.}

\item{renormalise}{Scale lambda to align with Campbell formula.}

\item{border}{Use translation correction? Default=1, yes. Only for cuboidal windows.}
}
\description{
Estimate the anisotropic pair correlation function (2d and 3d), as defined in f 1991, f. 5.2-5.3.
}
\details{
The antipode symmetry makes it necessary to compute only on one half of the circle/sphere.
}
