% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf_anisotropic.R
\name{pcf_anisotropic_old}
\alias{pcf_anisotropic_old}
\title{Anisotropic pair correlation function, old version}
\usage{
pcf_anisotropic_old(x, r, theta, h, stoyan = 0.15,
  correction = "translation", n_dir = 7)
}
\arguments{
\item{x}{pp, list with $x~coordinates $bbox~bounding box}

\item{r}{radius vector at which to estimate}

\item{theta}{vector or list of vectors for for angles, see details.}

\item{h}{widths of epanechnicov kernels, vector of two values, one for ranges and and one for angles.}

\item{stoyan}{If h not given, use h=c( stoyan/lambda^(1/dim), stoyan*pi) (cf. \code{pcf.ppp}-function in \code{spatstat}-package).}

\item{correction}{"none" or translation. Translation only for rectangle box.}

\item{n_dir}{Angle grid resolution (in case theta not given)

The antipode symmetry makes it necessary to compute only on one half of the circle/sphere.}
}
\description{
Estimate the anisotropic pair correlation function (2d and 3d), as defined in f 1991, f. 5.2-5.3.
}
