% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf_directional.R
\name{pcf_directional}
\alias{pcf_directional}
\title{Directed pcf function}
\usage{
pcf_directional(x, u, epsilon, r, ..., cylindrical = FALSE)
}
\arguments{
\item{x}{pp, spatstat's ppp-object, or a coordinate matrix, or a list with $x~coordinates $bbox~bounding box}

\item{u}{unit direction(s), if many then one direction per row.}

\item{epsilon}{Central angle for the directed cone (total angle is 2 epsilon)}

\item{r}{radius vector at which to evaluate K}

\item{...}{Passed on to \code{\link{pcf_anin}}}

\item{cylindrical}{If TRUE, compute the cylindrical version using \code{\link{pcf_anin_cylinder}}}
}
\value{
Returns a dataframe.
}
\description{
We use translation edge correction if we have an axis-oriented bounding box. Otherwise minus-correction.
}
\details{
Compute the sector/cone/cylindrical pcf function. This version uses the more general anisotropic-inhomonogeneous \link{pcf_anin_conical} and \code{\link{pcf_anin_cylinder}}, by setting the intensity = constant. See there for further parameters, especially kernel smoothing options.
}
\examples{

x <- matrix(runif(300), ncol=2)
k <- pcf_directional(x)
plot(k, rmax = 0.1)
}
