% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf_anisotropic_units.R
\name{pcf_directions}
\alias{pcf_directions}
\title{Anisotropic pair correlation function, direction vector formulation}
\usage{
pcf_directions(
  x,
  directions,
  h,
  f = 0.15,
  correction = "translation",
  n_dir,
  r
)
}
\arguments{
\item{x}{pp, list with $x~coordinates $bbox~bounding box}

\item{directions}{Matrix of direction vectors}

\item{h}{widths of epanechnicov kernels, vector of two values, for ranges and angles.}

\item{f}{If h not given, use h=f/lambda^(1/dim) for range and h=f*pi}

\item{correction}{"none" or translation. Translation only for rectangle box.}

\item{n_dir}{Direction vector grid resolution (if units not given). See Details.}

\item{r}{In case directions not given, use these lengths for generated directions.}
}
\description{
Estimate the anisotropic pair correlation function (2d and 3d), as defined in Stoyan 1991, f. 5.2-5.3.
This one doesn't use spherical coordinates but direction vectors.
}
\details{
For example, in 2D the polar vector (r, theta) is now a vector (r*cos(theta), r*sin(theta)).

For 2d, we use a regular grid of n_dir steps on [0,pi], leaving out pi due to antipode-symmetry to 0. 
n_dir=7 is the default.

For 3d we use a triangular grid on the expanding sphere. This is generated by
subdiving a regular icosahedron n_dir times. n_dir=2 is the default.
}
