% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf_sector.R
\name{pcf_sector}
\alias{pcf_sector}
\title{Sector pair correlation function}
\usage{
pcf_sector(
  x,
  r,
  theta,
  unit,
  epsilon,
  h,
  stoyan = 0.15,
  correction = "translation",
  n_dir = 12
)
}
\arguments{
\item{x}{pp, list with $x~coordinates $bbox~bounding box}

\item{r}{Ranges}

\item{theta}{Direction sector central angles, list of ncol(x$x)-components.}

\item{unit}{Alternative to theta, a matrix of direction vectors.}

\item{epsilon}{Direction sector width angles}

\item{h}{width of epanechnicov kernel}

\item{stoyan}{if h not given, use h=stoyan/sqrt(lambda)}

\item{correction}{"none" or translation. Translation only for rectangle box.}

\item{n_dir}{If unit missing, compute on a grid with resolution n_dir}
}
\description{
[DEFUNC] This is the same as "sector pair correlation function" in Stoyan 1991.
Direction defined with a unit vector. Estimated in a cone of angle theta and different radii.
}
\details{
Otherwise same as default pcf in spatstat, but now with an angle.
}
