% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodogram.R
\name{periodogram}
\alias{periodogram}
\title{Bartlett's periodogram}
\usage{
periodogram(x, p, pgrid, std = TRUE, ...)
}
\arguments{
\item{x}{point pattern, list of $x and $bbox}

\item{p}{number of frequencies, -p:p x dim grid}

\item{pgrid}{the grid of integer spectral frequencies}

\item{std}{Standardize coordinates?}

\item{...}{ignored

Omega=0 value is set to NA, and returned as $zerov in the return object.}
}
\description{
Bartlett's periodogram
}

