% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodogram.R
\name{periodogram_MR}
\alias{periodogram_MR}
\title{Periodogram as in Mugglestone and Renshaw 2001}
\usage{
periodogram_MR(x, ...)
}
\arguments{
\item{x}{point pattern, list of $x and $bbox}

\item{...}{ignored}
}
\description{
Periodogram as in Mugglestone and Renshaw 2001
}
