% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ellipsoid_methods.R
\name{persp.ellipsoid}
\alias{persp.ellipsoid}
\title{Perspective plot of an ellipsoid in 3d}
\usage{
\method{persp}{ellipsoid}(
  x,
  add = FALSE,
  theta = 25,
  phi = 30,
  expand = 0.9,
  triptych = FALSE,
  pmat,
  ...
)
}
\arguments{
\item{x}{ellipsoid-object}

\item{add}{Add to a scene?}

\item{theta}{Camera rotation, see persp.default}

\item{phi}{Camera rotation, see persp.default}

\item{expand}{zoom}

\item{triptych}{plot from all three main axes?}

\item{pmat}{optional projection matrix (e.g. from 2d persp)}
}
\description{
Try to make a pretty perspective plot of an 3D ellipsoid.
}
