% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pKolmogorov.R
\name{pKolmogorov}
\alias{pKolmogorov}
\title{Limiting distribution of the Kolmogorov-Smirnov test statistic}
\usage{
pKolmogorov(x, upto = 999)
}
\description{
Limiting distribution of the Kolmogorov-Smirnov test statistic
}
\details{
Will return approximately lim P(sqrt(n)Dn < x) .
}

