% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani-compression-profile.R
\name{plot.anisotropyprofile}
\alias{plot.anisotropyprofile}
\title{Plot method for "anisotropyprofile"}
\usage{
\method{plot}{anisotropyprofile}(x, ..., xlab = "first grid component",
  ylab = "anisotropy profile", scale = TRUE)
}
\description{
Plot method for "anisotropyprofile"
}
