% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbpoly.R
\name{plot.bbpoly}
\alias{plot.bbpoly}
\title{Plot bbpoly object}
\usage{
plot.bbpoly(
  x,
  normals = FALSE,
  faces = FALSE,
  ecol = 2,
  ncol = 4,
  add = TRUE,
  ...
)
}
\arguments{
\item{x}{bbpoly object}

\item{normals}{draw normals}

\item{faces}{draw faces}

\item{ecol}{edge color}

\item{ncol}{normal color}

\item{add}{add or not}

\item{...}{passsed to shade3d or lines (2d)}
}
\description{
Plot bbpoly object
}
