% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint-fry-ellipsoids.R
\name{plot.confint_fryellipsoids}
\alias{plot.confint_fryellipsoids}
\title{Plot confidence intervals}
\usage{
\method{plot}{confint_fryellipsoids}(x, ylim, ...)
}
\description{
Plot confidence intervals
}

