% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ellipsoid_methods.R
\name{plot.ellipsoid}
\alias{plot.ellipsoid}
\title{Plot an Ellipsoid Object}
\usage{
\method{plot}{ellipsoid}(
  x,
  add = TRUE,
  i = 0,
  levels = 1,
  res = 128,
  scale = 1,
  N = 2,
  xlab = "x",
  ylab = "y",
  ...
)
}
\arguments{
\item{x}{ellipsoid}

\item{add}{logical}

\item{i}{Normal 1,2 or 3. Plot intersection of 3d ellipsoid.}

\item{levels}{If i>0, draw this many levels from origin to tip of the axis}

\item{res}{2d resolution}

\item{scale}{1. use to enlarge or shrink}

\item{N}{2, iteration of refinement in 3D (resolution)}

\item{xlab}{axis-label}

\item{ylab}{axis-label}

\item{...}{passed on to lines or rgl::shade3d}
}
\description{
Internal ellipsoid plotter.
}
\details{
#' For 3D ellipsoid, if i=1,2 or 3, plot 2D intersection of 
plane:  i = 1 yz-plane; i = 2 xz-plane; i = 3 xy-plane.
}
