% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_method_f_cone.R
\name{plot.f_cone}
\alias{plot.f_cone}
\title{Plot a typical conical directed function of range}
\usage{
\method{plot}{f_cone}(x, r_scale = 1, rmax, ylim, legpos = "topright", lwd = 1, ...)
}
\arguments{
\item{x}{Output from some function}

\item{r_scale}{Plot with x-axis r*r_scale}

\item{rmax}{plot upto this range}

\item{ylim}{optional range for y-axis}

\item{legpos}{legend position}

\item{...}{passed on to plot}
}
\description{
Plot a typical conical directed function of range
}
\details{
Used by: nn_directional_distance_*
}
