% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fry-ellipsoids.R
\name{plot.fryellipsoids}
\alias{plot.fryellipsoids}
\title{plot method for fryellipsoids}
\usage{
\method{plot}{fryellipsoids}(
  x,
  ellipsoids = TRUE,
  used_points = TRUE,
  sectors = NULL,
  xlim = NULL,
  ylim = NULL,
  zoom = NULL,
  pch = 1,
  cex = 0.8,
  ...
)
}
\description{
plot method for fryellipsoids
}
