% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{plot.intensity_bw}
\alias{plot.intensity_bw}
\title{Quick plot for intensity bw object}
\usage{
plot.intensity_bw(x, ...)
}
\arguments{
\item{x}{output of \link{intensity_bandwidth_profile}}

\item{...}{passed on to plot.default}
}
\description{
Quick plot for intensity bw object
}
