% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kest_inan.R
\name{plot.K_anin}
\alias{plot.K_anin}
\title{Plot Kest_anin object}
\usage{
\method{plot}{K_anin}(x, r_scale = 1, rmax, legpos = "topleft", ...)
}
\arguments{
\item{x}{Output from Kest_anin}

\item{r_scale}{Plot with x-axis r*r_scale}

\item{rmax}{plot upto this range}

\item{legpos}{legend position}

\item{...}{passed on to plot}
}
\description{
Plot Kest_anin object
}

