% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markcorr_anin.R
\name{plot.markcorr_anin}
\alias{plot.markcorr_anin}
\title{Plot markcorr_anin object}
\usage{
plot.markcorr_anin(x, r_scale = 1, rmax, ylim, legpos = "topright", ...)
}
\arguments{
\item{x}{Output from markcor_anin}

\item{r_scale}{Plot with x-axis r*r_scale}

\item{rmax}{plot upto this range}

\item{ylim}{optional range for y-axis}

\item{legpos}{legend position}

\item{...}{passed on to plot}
}
\description{
Plot markcorr_anin object
}
