% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf_anin_fry.R
\name{plot.pcf_anin}
\alias{plot.pcf_anin}
\title{Plot pcf_anin object}
\usage{
\method{plot}{pcf_anin}(x, r_scale = 1, rmax, ylim, legpos = "topright", ...)
}
\arguments{
\item{x}{Output from pcf_anin or pcf_anin_cylinder}

\item{r_scale}{Plot with x-axis r*r_scale}

\item{rmax}{plot upto this range}

\item{ylim}{optional range for y-axis}

\item{legpos}{legend position}

\item{...}{passed on to plot}
}
\description{
Plot pcf_anin object
}
