% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.pcf_anisotropic}
\alias{plot.pcf_anisotropic}
\title{method plot for anisotropic pcf-function}
\usage{
\method{plot}{pcf_anisotropic}(x, ab, ...)
}
\description{
method plot for anisotropic pcf-function
}

