% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot2d.pcf_anisotropic}
\alias{plot2d.pcf_anisotropic}
\title{method plot2d for anisotropic pcf-function}
\usage{
plot2d.pcf_anisotropic(x, ab, ...)
}
\description{
method plot2d for anisotropic pcf-function
}
