% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbpoly.R
\name{poly2bbpoly}
\alias{poly2bbpoly}
\title{bbpoly from 2d polygon given by x and y coordinates}
\usage{
poly2bbpoly(x, y)
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}
}
\description{
bbpoly from 2d polygon given by x and y coordinates
}
\details{
Assuming an open polygon, meaning the last vertex is not the 
same as the first. Will check and adapt.
}

