% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani-compression-profile.R
\name{print.anisotropyprofile_summary}
\alias{print.anisotropyprofile_summary}
\title{Print method for "anisotropyprofile" summary}
\usage{
print.anisotropyprofile_summary(x, ...)
}
\description{
Print method for "anisotropyprofile" summary
}
