% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbpoly.R
\name{print.bbpoly}
\alias{print.bbpoly}
\title{print method for bbpoly, a cuboid made of quads}
\usage{
\method{print}{bbpoly}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{ignored}
}
\description{
print method for bbpoly, a cuboid made of quads
}
