% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbquad.R
\name{print.bbquad}
\alias{print.bbquad}
\title{print method for bbquad, a cuboid made of quads}
\usage{
\method{print}{bbquad}(x, ...)
}
\arguments{
\item{x}{bbquad}

\item{...}{ignored}
}
\description{
print method for bbquad, a cuboid made of quads
}
