% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/confint-fry-ellipsoids.R
\name{print.confint_fryellipsoids}
\alias{print.confint_fryellipsoids}
\title{Print method for confint_fryellipsoid}
\usage{
\method{print}{confint_fryellipsoids}(x, ...)
}
\description{
Print method for confint_fryellipsoid
}

