% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{print.intensity_bw}
\alias{print.intensity_bw}
\title{Print for intensity bw object}
\usage{
print.intensity_bw(x, ...)
}
\arguments{
\item{x}{output of \link{intensity_bandwidth_profile}}

\item{...}{ignored}
}
\description{
Print for intensity bw object
}
