% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ellipsoid_projections.R
\name{project_to_plane}
\alias{project_to_plane}
\title{Project a Point To Plane, 2D Outcome}
\usage{
project_to_plane(xyz, normal)
}
\arguments{
\item{xyz}{points to be projected, xyz-matrix}

\item{normal}{normal of the plane

We will simply rotate the points xyz and drop z-coordinate.
Rotation order: 1) Rz 2) Ry

This is not unique w.r.t. rotation around the normal, be aware.}
}
\description{
Plane defined using a point and a normal
}
