% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_sphere.R
\name{quaternion2EulerAngles}
\alias{quaternion2EulerAngles}
\title{Quaternion to Euler angles}
\usage{
quaternion2EulerAngles(q)
}
\arguments{
\item{q}{quartenion vector}
}
\description{
from euclideanspace.com
}
\details{
Convention: Euler is (Heading, Attitude, Bank) so that
Heading = rotation  around y, Attitude = rotation around z, 
Bank = rotation around x, in that order.
}
