% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_sphere.R
\name{quaternion2rotationMatrix}
\alias{quaternion2rotationMatrix}
\title{Quaternion to rotation matrix}
\usage{
quaternion2rotationMatrix(q)
}
\arguments{
\item{q}{quartenion vector}
}
\description{
Quaternion to rotation matrix
}
