% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gsector.R
\name{redenbach_border_term}
\alias{redenbach_border_term}
\title{The erosion of window with sector}
\usage{
redenbach_border_term(d, bbox, u, epsilon)
}
\description{
The erosion of window with sector
}
\details{
Called by 'nn_directional_distance_distribution' function.
}
