% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ellipsoid_rellipsoid.R
\name{rellipsoid_dev}
\alias{rellipsoid_dev}
\title{Sample points uniformly on the surface of an ellipsoid, 2D or 3D}
\usage{
rellipsoid_dev(
  n,
  axes = c(1, 1, 1),
  noise.sd = 0,
  R = NULL,
  method = 1,
  rej = FALSE
)
}
\arguments{
\item{axes}{vector giving the semi-axis lengths}

\item{noise.sd}{noise}

\item{R}{rotation matrix}
}
\description{
Sample points uniformly on the surface of an ellipsoid, 2D or 3D
}
