% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosenberg.R
\name{rosenberg}
\alias{rosenberg}
\title{Rosenberg Wavelet Transform}
\usage{
rosenberg(x)
}
\arguments{
\item{x}{pp, list with $x~coordinates $bbox~bounding box}
}
\description{
Estimate the wavelet variance suggested by Rosenberg 2004. 2D only.
}
