% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_sphere.R
\name{rotationMatrix2Euler}
\alias{rotationMatrix2Euler}
\title{Rotation matrix (prop or improp) to Euler axis-angle}
\usage{
rotationMatrix2Euler(R)
}
\arguments{
\item{R}{Rotation matrix}
}
\description{
Rotation matrix (prop or improp) to Euler axis-angle
}
