% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_sphere.R
\name{rotationMatrix2EulerAngles}
\alias{rotationMatrix2EulerAngles}
\title{Rotation matrix to Euler angles}
\usage{
rotationMatrix2EulerAngles(R)
}
\arguments{
\item{R}{rotation matrix}
}
\value{
(Heading, Attitude, Bank)
}
\description{
Rotation matrix to Euler angles
}
