% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_sphere.R
\name{rotationMatrix3}
\alias{rotationMatrix3}
\title{Product of elementary rotation matrices in 3D}
\usage{
rotationMatrix3(ax = 0, ay = 0, az = 0, order = c(3, 2, 1))
}
\arguments{
\item{order}{Order in which to to the product. Aimed for pre-product
Default: 3-2-1 i.e. around z, then around y, then around x.}
}
\description{
Product of elementary rotation matrices in 3D
}
