% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ellipsoid_methods.R
\name{sample_ellipse_beta}
\alias{sample_ellipse_beta}
\title{Sample from the OLS estimate of the beta parameters}
\usage{
sample_ellipse_beta(x, nsim = 100, tol = 0, maxiter = 500)
}
\arguments{
\item{x}{ellipsoid object}

\item{nsim}{number of simulations}

\item{tol}{tolerance when comparing to vector length 1}

\item{maxiter}{maximum number of iterations to try to get enough samples within tolerance}
}
\description{
Simulate the beta parameters, approximately normal conditional on ||beta||^2=1
}
