% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sectorplot.R
\name{sectorplot}
\alias{sectorplot}
\title{Plot a sector estimate}
\usage{
sectorplot(x, res = 4, col = rainbow, zlim, overlapfactor = 0.95,
  ...)
}
\arguments{
\item{x}{result object from one of the directed summaries}

\item{res}{smoothness of arc}

\item{col}{function to derive the colors, see values2colors}

\item{zlim}{passed to values2colors}

\item{overlapfactor}{<1 leaves a bit of space between sectors}

\item{...}{passed to plot, mostly the "main" parameter}
}
\description{
Plot a sector estimate
}
