% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbpoly-computations.R
\name{simplex_volume}
\alias{simplex_volume}
\title{Volume of a simplex}
\usage{
simplex_volume(x)
}
\arguments{
\item{x}{simplex, as a (d+1,d) matrix of vertices where d is dimension}
}
\description{
Volume of a simplex
}
\details{
If x is a list, we apply the function to each element.
}

