% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ellipsoid_OLS.R
\name{solve_S0}
\alias{solve_S0}
\title{Solve covariance matrix from unstable X}
\usage{
solve_S0(Y)
}
\arguments{
\item{Y}{The matrix, nrow < ncol}
}
\description{
Solve covariance matrix from unstable X
}
\details{
Solves the inverse of t(Y)%*%Y by slowly increasing diagonal (hedging).
}
