% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_periodogram.R
\name{spectrum}
\alias{spectrum}
\title{Mugglestone and Renshaw's polar spectrums}
\usage{
spectrum(x, std = TRUE, scale = TRUE, ...)
}
\arguments{
\item{x}{point pattern, list of $x and $bbox}

\item{std}{Standardize coordinates?}

\item{...}{ignored

(r,angle)=0 value is set to NA, and returned as $zerov in the return object.}

\item{r}{range frequencies}
}
\description{
Mugglestone and Renshaw's polar spectrums
}
