% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani-compression-profile.R
\name{summary.anisotropyprofile}
\alias{summary.anisotropyprofile}
\title{Summary for anisotropy profile}
\usage{
summary.anisotropyprofile(x, ...)
}
\description{
Summary for anisotropy profile
}
