% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnangle.R
\name{summary.nnangle}
\alias{summary.nnangle}
\title{Summary method for nearest neighbour angles}
\usage{
\method{summary}{nnangle}(x, ...)
}
\arguments{
\item{x}{object from nnangle-function}

\item{...}{passed on to ks.test for testing uniformity of the angles}
}
\description{
Summary method for nearest neighbour angles
}

