% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_directions.R
\name{theta_2_unit}
\alias{theta_2_unit}
\title{theta (angle) to unit vector, list format input}
\usage{
theta_2_unit(theta)
}
\arguments{
\item{list}{with 1 or 2 components, giving the angle(s)}
}
\description{
Convert from angle to unit vector
}
