% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbquad.R
\name{three_points_to_plane}
\alias{three_points_to_plane}
\title{Three points to a plane}
\usage{
three_points_to_plane(x)
}
\arguments{
\item{x}{vector of 3 coordinates}
}
\description{
Three points to a plane
}
