% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangulate-polygon-indices.R
\name{triangulate_indices_cohen_hickey}
\alias{triangulate_indices_cohen_hickey}
\title{Triangulate a polygon given by indices}
\usage{
triangulate_indices_cohen_hickey(edges)
}
\description{
Chooses the first vertex of the first edge as apex
}

