% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ellipsoid_projections.R
\name{triangulation_areas}
\alias{triangulation_areas}
\title{Compute the areas of polygons per 2D or 3D triangulation}
\usage{
triangulation_areas(tri)
}
\description{
Compute the areas of polygons per 2D or 3D triangulation
}
