% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_directions.R
\name{unit_2_theta}
\alias{unit_2_theta}
\title{unit vector to theta (angle) list format input}
\usage{
unit_2_theta(unit)
}
\arguments{
\item{unit}{vector(s as matrix, cols=dimensions)}
}
\description{
unit vector to theta (angle) list format input
}
