% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_sphere.R
\name{xyzrotate}
\alias{xyzrotate}
\title{rotate xyz coordinates}
\usage{
xyzrotate(xyz, ...)
}
\arguments{
\item{xyz}{coordinates}

\item{...}{passed to rotationMatrix3}
}
\description{
rotate xyz coordinates
}
