% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle-vector-methods.R
\name{angles.smooth}
\alias{angles.smooth}
\title{smoothen an histogram of angles}
\usage{
angles.smooth(binned, ...)
}
\arguments{
\item{binned}{binning of angles}

\item{...}{passed to loess}
}
\description{
smoothen an histogram of angles
}

