% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartletts_spectral_density.R
\name{bartletts_spectral_density}
\alias{bartletts_spectral_density}
\title{Bartlett's spectral density}
\usage{
bartletts_spectral_density(x, omega, ...)
}
\arguments{
\item{x}{point pattern, list of $x and $bbox}

\item{omega}{The frequencies}

\item{...}{ignored}
}
\description{
Periodogram estimator
}
\details{
We estimate the spectral density using a periodogram as suggested by Bartlett 1964,

\deqn{\mathcal{F}(\omega) = \lambda + \lambda^2\int_{R^d}[g(z)-1]e^{-i\omega^T z}dz}

where we assume that the process is stationary with intensity $\lambda$ and pair correlation function $g$. 
Isotropy is not assumed.

This function deliberately does not scale or assume a form for the frequencies (such as 2kpi/n, k integer), as 
there is no consensus on the best form. 

Additionally, no scaling or other transformation of the pattern is conducted. 

The frequencies that the spectrum is estimated are given by omega: 

1) If omega is a vector, we expand it to d-dimensional frequencies using expand.grid.

2) If omega is a column matrix of dimension m x d, each row is interpreted as a frequency. 

No edge correction is applied as none is known. The
}
