% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bbox_methods.R
\name{bbox_make}
\alias{bbox_make}
\title{compute the bounding box for coordinates}
\usage{
bbox_make(x)
}
\arguments{
\item{x}{coordinate matrix}
}
\description{
compute the bounding box for coordinates
}

