% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbpoly-computations.R
\name{bbpoly_intersection}
\alias{bbpoly_intersection}
\title{Intersection of two polytopes}
\usage{
bbpoly_intersection(a, b)
}
\arguments{
\item{a}{Polytope a}

\item{b}{Polytope b}
}
\description{
Compute "a intesection b"
}

