% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbquad.R
\name{bbquad_distance}
\alias{bbquad_distance}
\title{Distance from points to bbquad walls}
\usage{
bbquad_distance(x, b)
}
\arguments{
\item{x}{n x 3 matrix of locations}

\item{b}{quadrilateral bound box 

Note: Will check if points are inside or outside.}
}
\description{
Distance from points to bbquad walls
}
