% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_directions.R
\name{check_directions}
\alias{check_directions}
\title{Sector computation directions}
\usage{
check_directions(theta, epsilon, unit, dim = 2, antipode = FALSE,
  n_dir = 12)
}
\description{
Sector computation directions
}

