% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/confint-fry-ellipsoids.R
\name{confint.fryellipsoids}
\alias{confint.fryellipsoids}
\title{Confidence intervals for Fry ellipsoids}
\usage{
\method{confint}{fryellipsoids}(x, fun, ...)
}
\arguments{
\item{x}{Fitted Fry-ellipsoids object}

\item{fun}{contrast function, passed on to confint.ellipsoid}

\item{...}{Passed to confint.ellipsoid}
}
\description{
Confidence intervals for Fry ellipsoids
}
\details{
See the 'confint.ellipsoid' in 'ellipsoid' for details of ...
}

