% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{intensity_at_points}
\alias{intensity_at_points}
\title{Estimate intensity of point pattern at the points of the pattern}
\usage{
intensity_at_points(x, bw, n = 21, ...)
}
\description{
Estimate intensity of point pattern at the points of the pattern
}

