% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{intensity_at_points}
\alias{intensity_at_points}
\title{Estimate intensity of point pattern at the points of the pattern}
\usage{
intensity_at_points(x, bw, b = 1, ...)
}
\arguments{
\item{x}{points, list with $x and $bbox or just a matrix of coordinates.}

\item{bw}{bandwidth, Epanechnikov kernel domain [-bw,bw]}

\item{b}{Type of border correction estimate.}

\item{...}{ignored}
}
\description{
Estimate intensity of point pattern at the points of the pattern
}
\details{
Uses Epanechnikov kernel smoothing. Border corrections vary, try "b=1,2,3".
1: 2D exact, 3D box rectangle approximation; 2: 2D Biased box integral, 3D none; >3: fine grid sum with resolution b^dim (slow). b=0 means no correction.
}

