% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{intensity_at_points}
\alias{intensity_at_points}
\title{Kernel Estimate Intensity at Data Points}
\usage{
intensity_at_points(
  x,
  bw,
  kernel = "gaussian",
  border = "local",
  normalise = FALSE,
  loo = FALSE
)
}
\arguments{
\item{x}{point pattern}

\item{bw}{bandwidth. Gaussian sd=bw, Epanechnicov domain [-bw, bw].}

\item{kernel}{Either 'gaussian' or 'epanenchnikov' (partial matching)}

\item{border}{Border correction to apply.  One of 'none', 'local', 'global', 'toroidal'.}

\item{normalise}{renormalise so that inverse sum = volume}

\item{loo}{leave-one-out -estimate? Don't include point i for estimation of int(i)}
}
\description{
Kernel Estimate Intensity at Data Points
}
\details{
For border correction, the bounding box of the coordinates of x will be used, 
so at the moment it works only for axis-aligned cuboids.

The kernel will be a product of one dimensional kernels,
so Epanechnikov in many dimensions will not be truly isotropic but prefers diagonal directions.
}
