% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{intensity_somewhere}
\alias{intensity_somewhere}
\title{Estimate intensity of point pattern at given points}
\usage{
intensity_somewhere(x, loc, bw, n = 21, ...)
}
\description{
Estimate intensity of point pattern at given points
}

