% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{intensity_somewhere}
\alias{intensity_somewhere}
\title{Estimate intensity of point pattern at given locations}
\usage{
intensity_somewhere(x, loc, bw, b = 0, ...)
}
\arguments{
\item{x}{points, list with $x and $bbox or just a matrix of coordinates.}

\item{loc}{Matrix of locations to estimate the intensity at}

\item{bw}{bandwidth, Epanechnikov kernel domain [-bw,bw]}

\item{b}{Type of border correction estimate.}

\item{...}{ignored}
}
\description{
Estimate intensity of point pattern at given locations
}
\details{
Uses Epanechnikov kernel smoothing. Border corrections controlled by what b is:
0: 2D exact, not available for 3D atm; 
1: Box rectangle approximation; 
2: 2D Biased box integral, 3D none;
>3: fine grid sum with resolution b^dim (slow). 

if b < 0 no edge correction is applied.
}

