% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kdirectional.R
\docType{package}
\name{Kdirectional}
\alias{Kdirectional}
\alias{Kdirectional-package}
\title{Some tools for analysing the isotropy of spatial point patterns}
\description{
See the vignette for an example.
}
\examples{

\dontrun{
vignette("anisotropy-example")
}

}

