% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kest_directional.R
\name{Kest_directional}
\alias{Kest_directional}
\title{Directed K function}
\usage{
Kest_directional(x, u, epsilon, r, ..., cylindrical = FALSE)
}
\arguments{
\item{x}{pp, spatstat's ppp-object, or a coordinate matrix, or a list with $x~coordinates $bbox~bounding box}

\item{u}{unit direction(s), if many then one direction per row.}

\item{epsilon}{Central angle for the directed cone (total angle is 2 epsilon)}

\item{r}{radius vector at which to evaluate K}

\item{...}{Passed on to \code{\link{Kest_anin}}}

\item{cylindrical}{If TRUE, compute the cylindrical version using \code{\link{Kest_anin_cylinder}}}
}
\value{
Returns a dataframe.
}
\description{
We use translation edge correction if we have an axis-oriented bounding box. Otherwise minus-correction.
}
\details{
Compute the sector/cone/cylindrical K function. This version uses the more general anisotropic-inhomonogeneous \link{Kest_anin} and \code{\link{Kest_anin_cylinder}}, by setting the intensity = constant.
}
\examples{

x <- matrix(runif(200), ncol=2)
k <- Kest_directional(x)
plot(k, rmax = 0.1)

}
