% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnangle_in_cone.R
\name{nnangle_cone}
\alias{nnangle_cone}
\title{nearest neighbour angles and distances given that angles are in a cone}
\usage{
nnangle_cone(x, unit, theta, from, to, antipodal = FALSE)
}
\arguments{
\item{x}{matrix of coordinates}

\item{unit}{direction of the cone}

\item{theta}{angle or width of the cone}

\item{from}{indices from which to compute}

\item{to}{indices in which look for nearest neighbour}

\item{antipodal}{use antipodally symmetric vectors? default: FALSE

The values in 2d are (azimuth, distance).
The values in 3d are physical coordinate system (azimuth, inclination, distance)

The angle is inside the cone. NA if no neighbour is found.}
}
\description{
nearest neighbour angles and distances given that angles are in a cone
}
