% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf_anisotropic.R
\name{pcf_anisotropic}
\alias{pcf_anisotropic}
\title{Anisotropic pair correlation function}
\usage{
pcf_anisotropic(x, r, theta, h, f = 0.15, correction = "translation",
  n_dir = 7)
}
\arguments{
\item{x}{pp, list with $x~coordinates $bbox~bounding box}

\item{r}{radius vector at which to estimate}

\item{theta}{vector or list of vectors for for angles, see details.}

\item{h}{widths of epanechnicov kernels, vector of two values, for ranges and angles.}

\item{f}{If h not given, use h=c( f/lambda^(1/dim), h=f*pi). Same as 'stoyan' in spatstat's pcf.}

\item{correction}{"none" or translation. Translation only for rectangle box.}

\item{n_dir}{Angle grid resolution (in case theta not given)

The antipode symmetry makes it necessary to compute only on one half of the circle/sphere.}
}
\description{
Estimate the anisotropic pair correlation function (2d and 3d), as defined in f 1991, f. 5.2-5.3.
}

