% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint-fry-ellipsoids.R
\name{summary.confint_fryellipsoids}
\alias{summary.confint_fryellipsoids}
\title{Summary method for confint_fryellipsoid}
\usage{
\method{summary}{confint_fryellipsoids}(x, ...)
}
\description{
Summary method for confint_fryellipsoid
}
