% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint-fry-ellipsoids.R
\name{summary.confint_fryellipsoids}
\alias{summary.confint_fryellipsoids}
\title{Summary method for confint_fryellipsoid}
\usage{
summary.confint_fryellipsoids(object, ...)
}
\arguments{
\item{...}{ignored}

\item{Fitted}{Fry-ellipsoids object}
}
\description{
Summary method for confint_fryellipsoid
}
\details{
Extracts the CI-table.
}
