% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation_weights.R
\name{translation_weights}
\alias{translation_weights}
\title{Translation correction weights for point pattern}
\usage{
translation_weights(x, asMatrix = FALSE)
}
\arguments{
\item{x}{Pattern, $x coordinate matrix, $bbox bounding box.}
}
\description{
Defined in a bounding box.
}
