% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation_weights.R
\name{translation_weights}
\alias{translation_weights}
\title{Translation correction weights for point pattern}
\usage{
translation_weights(x, asMatrix = FALSE)
}
\arguments{
\item{x}{Pattern, $x coordinate matrix, $bbox bounding box.}

\item{asMatrix}{Return a square symmetric matrix? Otherwise only the upper triangle row by row.}
}
\value{
if asMatrix=FALSE, the vector gives the weights in order 
(1,2), (1,3), ..., (1,n), (2,3), (2,4), ... (n-2, n-1), (n-2,n), (n-1, n)
}
\description{
Defined in a bounding box.
}
