% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet-scale-angle-energy-density.R
\name{wavelet_saed}
\alias{wavelet_saed}
\title{Scale-angle-energy-density}
\usage{
wavelet_saed(x, theta = seq(0, pi, l = 30), shift_res = 20, scales)
}
\arguments{
\item{x}{point pattern, list of $x and $bbox}

\item{theta}{angles}

\item{shift_res}{resolution of the shift-grid}

\item{...}{ignored}
}
\description{
2D at the moment. Morlet anisotropic wavelet
}
